requirements.txt
mutagen==1.45.1
Pillow==8.2.0
pkg-resources==0.0.0
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.8.1
PySide6==6.0.3
shiboken6==6.0.3
